DuBoisPlots

Documentation for DuBoisPlots.

This package allows you to create beatiful plots in the style of W. E. B. DuBois effortlessy. For example,

using DuBoisPlots
using DataFrames
df = DataFrame(
        Year = [1790,1800,1810,1820,1830,1840,1850,1860,1870],
        Slaves = [92,88.5,86.5,87,86,87,88,89,0],
        Free =  [8.0, 11.5, 13.5, 13.0, 14.0, 13.0, 12.0, 11.0,100],
        )
Plate51(df, :Year; title="Proportion of Slaves")

Index

DuBoisPlots.Plate25

DuBoisPlots.Plate51

DuBoisPlots.Plate53

DuBoisPlots.Plate7

Automatic docs

DuBoisPlots.Plate25 — Function

Replicates Plate 25 (Spiral line chart).

You must pass a DataFrame with two columns. You must specify the column symbol for both columns.

df = DataFrame(Year = sort(rand(6)), Value = rand(6))
Plate25(data, :Year, :Value, "Title")

DuBoisPlots.Plate51 — Method

Replicates Plate 51 (Stacked Area Chart).

You must pass a DataFrame with three columns. You must specify the column symbol for the x-axis variable. Each row of the remaining variables must sum to 100.

df = DataFrame(t=[1,2],x1=[25,35],x2=[75,65])
Plate51(df,:t)

DuBoisPlots.Plate53 — Method

Replicates Plate 53 (Violin Chart).

This is a violin chart which is effectively made of two symmetric charts. The first column of the dataframe should be the categories that define those two plots. The second column of the dataframe should be the sub-categories within the plots – they will be the different colored regions within the chart area.

The remaining columns will be the values in each region. A number of things are important here. First, each row of these columns should sum to 100 or 1.0. Second, the way this chart works is that they are effectively horizontal stacked bar charts with mirrored x axes. There are as many bars as unique values in the sub-category column.

r = DataFrame(y_var = repeat(["A", "B", "C", "D","E","F"],2),
  cat_var = cat(repeat(["G"],6),repeat(["H"],6), dims = (1,1)),
  x1 = (rand(1:10, 12).+20),
  x2 = (rand(1:10, 12).+10),
  x3 = (rand(1:10, 12).+5),
  x4 = (rand(1:10, 12).+2))

  r.x5 = (100 .- (r.x1 + r.x2 + r.x3 + r.x4))

Plate53(r, :y_var, :cat_var, [:x1, :x2, :x3, :x4, :x5], "Auto", "Auto", title_1 = "Title",
  title_2 = "Small Title", subtitle = "Subtitle", bot_lab = "Label")

DuBoisPlots.Plate7 — Method

Replicates Plate 7 (Flipped Time Series).

You must pass a DataFrame with at least two columns. You must specify the column symbol for the x-axis variable.

df = DataFrame(t=[1,2],x1=[25,35],x2=[75,65])
Plate7(df,:t)