DuBoisPlots

Documentation for DuBoisPlots.

This package allows you to create beatiful plots in the style of W. E. B. DuBois effortlessy. For example,

using DuBoisPlots
using DataFrames
df = DataFrame(
        Year = [1790,1800,1810,1820,1830,1840,1850,1860,1870],
        Slaves = [92,88.5,86.5,87,86,87,88,89,0],
        Free =  [8.0, 11.5, 13.5, 13.0, 14.0, 13.0, 12.0, 11.0,100],
        )
Plate51(df, :Year; title="Proportion of Slaves")

Index

DuBoisPlots.Plate51

DuBoisPlots.Plate7

Automatic docs

DuBoisPlots.Plate51 — Method

Replicates Plate 51 (Stacked Area Chart).

You must pass a DataFrame with three columns. You must specify the column symbol for the x-axis variable. Each row of the remaining variables must sum to 100.

df = DataFrame(t=[1,2],x1=[25,35],x2=[75,65])
Plate51(df,:t)

DuBoisPlots.Plate7 — Method

Replicates Plate 7 (Flipped Time Series).

You must pass a DataFrame with at least two columns. You must specify the column symbol for the x-axis variable.

df = DataFrame(t=[1,2],x1=[25,35],x2=[75,65])
Plate7(df,:t)